#6: refactoring
